unbundle parseJsx
