+++ packages/core/src/parsers/jsx/context.ts
@@ -1,0 +1,57 @@
+import * as babel from '@babel/core';
+import { traceReferenceToModulePath } from '../../helpers/trace-reference-to-module-path';
+import traverse from 'traverse';
+import { createMitosisNode } from '../../helpers/create-mitosis-node';
+import { isMitosisNode } from '../../helpers/is-mitosis-node';
+import { MitosisComponent } from '../../types/mitosis-component';
+import { parseStateObject } from './state';
+
+const jsxPlugin = require('@babel/plugin-syntax-jsx');
+const tsPreset = require('@babel/preset-typescript');
+
+const { types } = babel;
+
+type Context = {
+  // Babel has other context
+  builder: {
+    component: MitosisComponent;
+  };
+};
+
+const expressionToNode = (str: string) => {
+  const code = `export default ${str}`;
+  return (
+    (babel.parse(code) as babel.types.File).program.body[0] as babel.types.ExportDefaultDeclaration
+  ).declaration;
+};
+
+/**
+ * Convert <Context.Provider /> to hooks formats by mutating the
+ * MitosisComponent tree
+ */
+export function extractContextComponents(json: MitosisComponent) {
+  traverse(json).forEach(function (item) {
+    if (isMitosisNode(item)) {
+      if (item.name.endsWith('.Provider')) {
+        const value = item.bindings?.value?.code;
+        const name = item.name.split('.')[0];
+        const refPath = traceReferenceToModulePath(json.imports, name)!;
+        json.context.set[refPath] = {
+          name,
+          value: value
+            ? parseStateObject(expressionToNode(value) as babel.types.ObjectExpression)
+            : undefined,
+        };
+
+        this.update(
+          createMitosisNode({
+            name: 'Fragment',
+            children: item.children,
+          }),
+        );
+      }
+      // TODO: maybe support Context.Consumer:
+      // if (item.name.endsWith('.Consumer')) { ... }
+    }
+  });
+}
