+++ packages/core/src/parsers/jsx/index.ts
@@ -1,1 +1,4 @@
 export * from './jsx';
+export { selfClosingTags } from './helpers';
+export { METADATA_HOOK_NAME } from './metadata';
+export { parseStateObject, createFunctionStringLiteralObjectProperty } from './state';
