+++ packages/core/src/parsers/jsx/types.ts
@@ -1,0 +1,5 @@
+export type ParseMitosisOptions = {
+  format: 'react' | 'simple';
+  jsonHookNames?: string[];
+  compileAwayPackages?: string[];
+};
